move cast to conversion_funcs
