+++ native/spark-expr/src/conversion_funcs/mod.rs
@@ -1,0 +1,3 @@
+pub mod cast;
+
+pub use cast::{spark_cast, Cast, SparkCastOptions};
